refactor: deduplicate UnifiedSuggestion type to web app types
- Create apps/web/src/types/suggestion.ts with unified type
- Remove duplicate interface definitions from 7 workbench components
- Export type from @/types for consistent usage across web app
- Re-export from hooks/index.ts for backwards compatibility
