+++ apps/web/src/types/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Shared Frontend Types
+ */
+
+export * from './suggestion';
